Crate pid_control

Software PID controller

This crate implements a PID controller. It has seen some amount of real-world usage driving 100+ kW electrical motors, but has not been tested to death. Use with caution (but do use it and file bug reports!).

Any change in behaviour that may make calculations behave differently will result in a major version upgrade; your tunings are safe as long as you stay on the same major version.

Owes a great debt to:

https://en.wikipedia.org/wiki/PID_controller

http://www.embedded.com/design/prototyping-and-development/4211211/PID-without-a-PhD

http://brettbeauregard.com/blog/2011/04/improving-the-beginners-pid-introduction/

Modules

util
Utility module.

Structs

PIDController
PID Controller.

Enums

DerivativeMode
PID controller derivative modes.

Traits

Controller
A generic controller interface.