Struct pid_control::PIDController

pub struct PIDController {
    pub p_gain: f64,
    pub i_gain: f64,
    pub d_gain: f64,
    pub i_min: f64,
    pub i_max: f64,
    pub out_min: f64,
    pub out_max: f64,
    pub d_mode: DerivativeMode,
    // some fields omitted
}

PID Controller.

A PID controller, supporting the Controller interface. Any public values are safe to modify while in operation.

p_gain, i_gain and d_gain are the respective gain values. The controlller internally stores an already adjusted integral, making it safe to alter the i_gain - it will not result in an immediate large jump in controller output.

i_min and i_max are the limits for the internal integral storage. Similarly, out_min and out_max clip the output value to an acceptable range of values. By default, all limits are set to +/- infinity.

d_mode The DerivativeMode, the default is OnMeasurement.

Fields

p_gain

Proportional gain

i_gain

Integral gain

d_gain

Differential gain,

i_min

i_max

out_min

out_max

d_mode

Methods

impl PIDController

fn new(p_gain: f64, i_gain: f64, d_gain: f64) -> PIDController

Creates a new PID Controller.

fn set_limits(&mut self, min: f64, max: f64)

Convenience function to set i_min/i_max and out_min/out_max to the same values simultaneously.